Swift:  Bug and style fixes (#220)
* Updated Makefile to use clean

Calling `make clean` causes a hard reset - obliterating the .build folder.

* Calling make commands is now prettier

Added `@` before command calls in each recipe to silence the command in
the cli, but still get the ouput of the command.  Looks so good,
giggity.

* Fixed switch case "="

Somehow "=" became .equal again instead of .assign.

* Removed Foundation

Foundation was not being used in any of the files and is a huge
library.  The import was removed.

* Token is now Equatable

Equatable protocol does not require a static == function be defined.

* Updated swift style and fixed out of bounds error

had to go back to re-commit every from an earlier commit.
- consolidated endIndex and isNumber/isValidIdentifier checks into one line. This then allowed to eliminate one of the returns.
- Added currentChar as computed property
- Removed initializer, set currentIndex to input.startIndex. The default initializer is then used which requires input in the argument list.
- String out of bounds error occurring again. Rearranged nextToken such that whitespace is passed over first (checking for endIndex with each loop to avoid a fault), and then doing another guard check on the index before proceeding.
- "=" in switch statement of nextToken now appropriately returns assign
- sqirly are now actually squirly
- Added Equitable conformance to Token which eliminated the need for the static == definition
- Updated the expectedTokens list to reflect to changes made to the Token type
- Updated the XCTAsserts in each test - only one assert is now needed with the new Token type.

* Update Token names

- updated equals and notEquals to equal and notEqual as before my
  changes disappeared into the ether.

* Update file naming to current convention

File names in Swift are generally named after the main structure or
class they contain.  PascalCase is usually used for structures and
classes and this is extended to the file name.  The exception is
main.swift which is left lowercase for historic reasons.

* Comment removed

Comment marks placed when testing were removed.

* Updated main.swift so now it runs

Calling swift run Lexer now calls the main function.

* Updated testNextToken, fixed off-by-one error

`testNextToken()` now checks for the code snippet from the book.  When
testing, an off-by-one error occurred after every `.ident` or `.int`.
Offset less 1 fixed the error.

* Updated file names

Prior commmit was not updating file names as expected due to an
idiosyncrasy of macOS which is case sensative, but git, by default is
case insensative.

* Extended Token coverage

Tokens slash, bang, minus, less than, greater than, true, false, if,
else, return, equal, not equal are now available.  The Lexer test was
updated to include these tokens.

Added some additional output in Makefile.

* Extended Token coverage

Tokens slash, bang, minus, less than, greater than, true, false, if,
else, return, equal, not equal are now available.  The Lexer test was
updated to include these tokens.

Added some additional output in Makefile.

* Initializer added

Original repo author prefers and would like to use either init or a
builder function to avoid use of a lazy var.  By using both an
initializer and non-lazy currentIndex, the currentChar implementation no
longer requires a mutating getter.

* Added keywords

Moved keywords into a dictionary which is checked when looking for an
ident.  Provided default keywords for initialization.

* Added default keywords static property to Token

---------

Co-authored-by: Myles La Verne Schultz <myleslschultz@icloud.com>